Separate things out and first pass at naming them
